libssl: don't declare, define and call pushsig() and popsig().
This is a workaround for a weird bug in the toolchain where elf2flt creates a
binary that cannot be loaded by the binfmt_flat loader in the kernel. This
happens to any program trying to use libssl.

BINFMT_FLAT: reloc outside program 0xb8050 (0 - 0xb79e0/0x93160), killing crypttest!
SIGSEGV